blobheader now uses a 32bit blog size, not 64bit

+++ src/3blob_manager/blob_manager_disk.h
@@ -144,7 +144,7 @@ class DiskBlobManager : public BlobManager
                     ByteArray *arena);
 
     // retrieves the size of a blob
-    virtual uint64_t do_get_blob_size(Context *context, uint64_t blobid);
+    virtual uint32_t do_get_blob_size(Context *context, uint64_t blobid);
 
     // overwrite an existing blob
     //
